requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.7
PyMySQL==1.1.1
cryptography==42.0.8
Werkzeug==3.0.3
python-dotenv==1.0.1
requests==2.32.3
Pillow==10.4.0
WTForms==3.1.2
email-validator==2.2.0
